v3.1.1 — make archive-watchlist pip-installable as console script
Move archive_watchlist.py into the iagitup package and register it as
an `archive-watchlist` console entry point so both commands are available
after `pip install iagitup`. Root script kept as thin backward-compat wrapper.
Update README, tests, and bump version to 3.1.1.
