+++ iagitup/iagitup.py
@@ -5,7 +5,7 @@ __author__ = "Giovanni Damiola"
 __copyright__ = "Copyright 2018-2026, Giovanni Damiola"
 __main_name__ = "iagitup"
 __license__ = "GPLv3"
-__version__ = "v3.1.0"
+__version__ = "v3.1.1"
 
 import configparser
 import logging
