+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "iagitup"
-version = "3.1.0"
+version = "3.1.1"
 description = "Archive GitHub repositories to the Internet Archive."
 readme = { file = "README.md", content-type = "text/markdown" }
 license = { text = "GPL-3.0-or-later" }
@@ -37,6 +37,7 @@ Repository = "https://github.com/gdamdam/iagitup"
 
 [project.scripts]
 iagitup = "iagitup.__main__:main"
+archive-watchlist = "iagitup.archive_watchlist:main"
 
 [tool.setuptools.packages.find]
 include = ["iagitup*"]
